requirements.txt
flask
altair
vega_datasets
Flask-WTF
psycopg2-binary
Flask-SQLAlchemy
gunicorn
python-dotenv
Flask-Migrate
matplotlib
azure-data-tables
python-dotenv